feat: v3.0 api redesign
This commit implements a redesign of the Fernet base api.

Focus on the redesign was to remove unneded abstractions, name methods and parameters more appropiately, improve unit testing and support PHP 8.0.

BREAKING CHANGE: Unfortunately, this release breaks BC so make all this improvements. However, public api surface of this library is so small that migration should be
effortless.

Version 2.x will still be mantained for security patches and bugfixes only.